Refactored and added app.js to separate server initalization
